pyproject.toml
[tool.poetry]
name = "monitoring service"
version = "0.1.0"
description = "Test for some employer"
authors = ["Sergey Guzun <dev.py.serg@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.95.1"
pydantic = "1.10.7"
starlette = "0.26.1"
uvicorn = "0.21.1"
alembic = "1.10.3"
python-dotenv = "1.0.0"
psycopg2-binary = "^2.9.4"
asyncpg = "^0.27.0"
sqlalchemy = "^2.0.10"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.1"
pytest = "^7.2.2"
black = "^23.3.0"
isort = "^5.12.0"
pytest-asyncio = "^0.21.0"
httpx = "0.26"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "/src/tests",
]

[tool.black]
line_length = 80

[tool.isort]
line_length = 80
profile = "black"
known_first_party = ["api", "infrastructure", "domain"]
